requirements.txt
selenium==4.12.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.4
webdriver-manager==4.0.1
